refactor(suite-native): extract RootStack into its own file
The ~170-line block of <Stack.Screen> declarations was the bulk of
_layout.tsx, mixing the route inventory with provider composition and
app-shell concerns. Extracting it makes _layout.tsx about half its
previous size and gives the route catalog a discoverable home anyone
adding or removing a route can navigate to directly.

slideFromBottomOptions moves with RootStack since it's only used there.
